Create an interface for externalVocabularyService
