+++ src/test/java/cz/cvut/kbss/termit/rest/VocabularyControllerSecurityTest.java
@@ -34,7 +34,7 @@ import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
 import static cz.cvut.kbss.termit.environment.Generator.generateVocabulary;
-import cz.cvut.kbss.termit.service.business.ExternalVocabularyService;
+import cz.cvut.kbss.termit.service.business.SparqlExternalVocabularyService;
 import static cz.cvut.kbss.termit.util.Constants.REST_MAPPING_PATH;
 import static org.mockito.Mockito.any;
 import static org.mockito.Mockito.never;
@@ -55,7 +55,7 @@ class VocabularyControllerSecurityTest extends BaseControllerTestRunner {
     private IdentifierResolver idResolverMock;
 
     @MockBean
-    private ExternalVocabularyService externalVocabularyService;
+    private SparqlExternalVocabularyService externalVocabularyService;
 
     @Autowired
     private Configuration configuration;
